+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # NEXT RELEASE
-* Rename BooleanFilter updateValidOptions() to updateOption()
+* Rename `BooleanFilter`'s `updateValidOptions()` to `updateOption()`
 * Completely reset scorers when `reset()` is called
 * Suppress foreign language and muted keywrod trending tags completely